+++ quill-core/src/main/scala/io/getquill/dsl/QueryDsl.scala
@@ -1,16 +1,17 @@
 package io.getquill.dsl
 
 import scala.language.experimental.macros
-import scala.reflect.ClassTag
 
 import io.getquill.quotation.NonQuotedException
 import scala.annotation.compileTimeOnly
 
 private[dsl] trait QueryDsl {
-  this: CoreDsl =>
+  dsl: CoreDsl =>
+
+  def query[T]: EntityQuery[T] = macro QueryDslMacro.expandEntity[T]
 
   @compileTimeOnly(NonQuotedException.message)
-  def query[T](implicit ct: ClassTag[T]): EntityQuery[T] = NonQuotedException()
+  def querySchema[T](entity: String, columns: (T => (Any, String))*): EntityQuery[T] = NonQuotedException()
 
   sealed trait Query[+T] {
 
@@ -62,36 +63,29 @@ private[dsl] trait QueryDsl {
   sealed trait EntityQuery[T]
     extends Query[T] {
 
-    def schema(f: Schema[T] => Schema[T]): EntityQuery[T]
-
     override def withFilter(f: T => Boolean): EntityQuery[T]
     override def filter(f: T => Boolean): EntityQuery[T]
     override def map[R](f: T => R): EntityQuery[R]
 
-    def insert(value: T): Insert[T] = macro macroz.DslMacro.expandInsert[T]
+    def insert(value: T): Insert[T] = macro QueryDslMacro.expandInsert[T]
     def insert(f: (T => (Any, Any)), f2: (T => (Any, Any))*): Insert[T]
 
-    def update(value: T): Update[T] = macro macroz.DslMacro.expandUpdate[T]
+    def update(value: T): Update[T] = macro QueryDslMacro.expandUpdate[T]
     def update(f: (T => (Any, Any)), f2: (T => (Any, Any))*): Update[T]
 
     def delete: Delete[T]
   }
 
-  sealed trait Schema[T] {
-    def entity(alias: String): Schema[T]
-    def columns(propertyAlias: (T => (Any, String))*): Schema[T]
-  }
+  sealed trait Action[E]
 
-  sealed trait Action[Entity]
-
-  sealed trait Insert[Entity] extends Action[Entity] {
+  sealed trait Insert[E] extends Action[E] {
     @compileTimeOnly(NonQuotedException.message)
-    def returning[R](f: Entity => R): ActionReturning[Entity, R] = NonQuotedException()
+    def returning[R](f: E => R): ActionReturning[E, R] = NonQuotedException()
   }
 
-  sealed trait ActionReturning[Entity, Output] extends Action[Entity]
-  sealed trait Update[Entity] extends Action[Entity]
-  sealed trait Delete[Entity] extends Action[Entity]
+  sealed trait ActionReturning[E, Output] extends Action[E]
+  sealed trait Update[E] extends Action[E]
+  sealed trait Delete[E] extends Action[E]
 
   sealed trait BatchAction[+A <: Action[_]]
 }