+++ quill-core/src/test/scala/io/getquill/context/QueryMacroSpec.scala
@@ -7,12 +7,19 @@ import io.getquill.testContext._
 
 class QueryMacroSpec extends Spec {
 
+  "runs query with nested optional case class" in {
+    val q = quote {
+      qr1.leftJoin(qr2).on((a, b) => a.i == b.i)
+    }
+    testContext.run(q)
+  }
+
   "runs query without liftings" in {
     val q = quote {
       qr1.map(t => t.i)
     }
     testContext.run(q).string mustEqual
-      "query[TestEntity].map(t => t.i)"
+      """querySchema("TestEntity").map(t => t.i)"""
   }
 
   "runs query with liftings" - {
@@ -22,7 +29,7 @@ class QueryMacroSpec extends Spec {
           qr1.filter(t => t.i == lift(1)).map(t => t.i)
         }
         val r = testContext.run(q)
-        r.string mustEqual "query[TestEntity].filter(t => t.i == ?).map(t => t.i)"
+        r.string mustEqual """querySchema("TestEntity").filter(t => t.i == ?).map(t => t.i)"""
         r.prepareRow mustEqual Row(1)
       }
       "two" in {
@@ -30,7 +37,7 @@ class QueryMacroSpec extends Spec {
           qr1.filter(t => t.i == lift(1) && t.s == lift("a")).map(t => t.i)
         }
         val r = testContext.run(q)
-        r.string mustEqual "query[TestEntity].filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"
+        r.string mustEqual """querySchema("TestEntity").filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"""
         r.prepareRow mustEqual Row(1, "a")
       }
       "nested" in {
@@ -41,7 +48,7 @@ class QueryMacroSpec extends Spec {
           qr1.filter(t => c(t) && t.s == lift("a")).map(t => t.i)
         }
         val r = testContext.run(q)
-        r.string mustEqual "query[TestEntity].filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"
+        r.string mustEqual """querySchema("TestEntity").filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"""
         r.prepareRow mustEqual Row(1, "a")
       }
     }
@@ -51,7 +58,7 @@ class QueryMacroSpec extends Spec {
           qr1.filter(t => t.i == lift(1)).map(t => t.i)
         }
         val r = testContext.run(q.dynamic)
-        r.string mustEqual "query[TestEntity].filter(t => t.i == ?).map(t => t.i)"
+        r.string mustEqual """querySchema("TestEntity").filter(t => t.i == ?).map(t => t.i)"""
         r.prepareRow mustEqual Row(1)
       }
       "two" in {
@@ -59,7 +66,7 @@ class QueryMacroSpec extends Spec {
           qr1.filter(t => t.i == lift(1) && t.s == lift("a")).map(t => t.i)
         }
         val r = testContext.run(q.dynamic)
-        r.string mustEqual "query[TestEntity].filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"
+        r.string mustEqual """querySchema("TestEntity").filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"""
         r.prepareRow mustEqual Row(1, "a")
       }
       "nested" in {
@@ -70,7 +77,7 @@ class QueryMacroSpec extends Spec {
           qr1.filter(t => c(t) && t.s == lift("a")).map(t => t.i)
         }
         val r = testContext.run(q.dynamic)
-        r.string mustEqual "query[TestEntity].filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"
+        r.string mustEqual """querySchema("TestEntity").filter(t => (t.i == ?) && (t.s == ?)).map(t => t.i)"""
         r.prepareRow mustEqual Row(1, "a")
       }
     }