+++ quill-jdbc/src/test/scala/io/getquill/context/jdbc/postgres/JdbcEncodingSpec.scala
@@ -28,13 +28,13 @@ class JdbcEncodingSpec extends EncodingSpec {
 
   "returning custom type" in {
     implicit val uuidDecoder: Decoder[UUID] =
-      decoder[UUID] { row => index => UUID.fromString(row.getObject(index).toString)
-      }
+      decoder[UUID]({ row => index => UUID.fromString(row.getObject(index).toString)
+      }, java.sql.Types.OTHER)
     implicit val uuidEncoder: Encoder[UUID] =
       encoder[UUID](row => (idx, uuid) =>
         row.setObject(idx, uuid, Types.OTHER), Types.OTHER)
 
-    val uuid = testContext.run(insertBarCode(lift(barCodeEntry))).get
+    val uuid = testContext.run(insertBarCode.apply(lift(barCodeEntry))).get
     val (barCode :: Nil) = testContext.run(findBarCodeByUuid(uuid))
 
     verifyBarcode(barCode)